Add more explicit casts. Delete unused variables. Add assert to id_read for unre cognized command. Fix unitialized variables.

+++ 3B2/3b2_mmu.h
@@ -278,7 +278,7 @@
 /* Fault codes */
 #define MMU_FAULT(f) {                                      \
         if (fc) {                                           \
-            mmu_state.fcode = ((r_acc<<7)|(CPU_CM<<5)|f);   \
+            mmu_state.fcode = ((((uint32)r_acc)<<7)|(((uint32)(CPU_CM))<<5)|f);   \
             mmu_state.faddr = va;                           \
         }                                                   \
     }
